Modularize session persistence queries
Separate session reads and writes, participant mappings, enhanced notes, event-backed creation, deletion, and public record types behind the existing query facade without changing SQL or behavior.
