+++ apps/desktop/src/session/queries/types.ts
@@ -1,0 +1,44 @@
+export type SessionRecord = {
+  id: string;
+  user_id: string;
+  created_at: string;
+  folder_id: string;
+  event_json: string;
+  title: string;
+  raw_md: string;
+};
+
+export type SessionChanges = Partial<
+  Pick<
+    SessionRecord,
+    "created_at" | "event_json" | "folder_id" | "raw_md" | "title"
+  >
+>;
+
+export type SessionSummaryRecord = {
+  id: string;
+  title: string;
+  created_at: string;
+};
+
+export type EnhancedNoteRecord = {
+  id: string;
+  sessionId: string;
+  title: string;
+  content: string;
+  templateId: string;
+  position: number;
+};
+
+export type SessionParticipantRecord = {
+  id: string;
+  sessionId: string;
+  humanId: string;
+  source: string;
+  name: string;
+  email: string;
+  jobTitle: string;
+  linkedinUsername: string;
+  organizationId: string;
+  organizationName: string;
+};
